refactor: move app to separate file
